requirements.txt
aiobotocore==2.23.2
aiohappyeyeballs==2.4.6
python-dotenv
aiohttp==3.11.12
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.5.3
arrow==1.3.0
asttokens==3.0.0
async-timeout==5.0.1
asyncssh==2.19.0
atpublic==5.1
attrs==25.1.0
awscli
backoff==2.2.1
billiard==4.2.1
binaryornot==0.4.4
blinker==1.9.0
boto3
botocore
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
configobj==5.0.9
contourpy==1.3.1
cookiecutter==2.6.0
cryptography==44.0.1
cycler==0.12.1
dacite==1.6.0
dagshub==0.5.4
dagshub-annotation-converter==0.1.3
databricks-sdk==0.44.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.2.18
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docutils==0.19
dpath==2.2.0
dulwich==0.22.7
dvc==3.59.0
dvc-data==3.16.9
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
filelock==3.17.0
Flask==3.1.0
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
funcy==2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
gql==3.5.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.1.1
gto==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
ipykernel==6.30.1
ipython==8.37.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
kombu==5.4.2
lxml==5.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mglearn==0.2.0
mlflow==2.20.2
mlflow-skinny==2.20.2
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
omegaconf==2.3.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
optuna==4.4.0
orjson==3.10.15
packaging
pandas==2.2.3
parso==0.8.5
pathspec==0.12.1
pathvalidate==3.2.3
pillow==11.1.0
pipx==1.7.1
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.2.1
protobuf==5.29.3
psutil==7.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pygit2==1.17.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.7.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==2025.2.0
s3transfer==0.13.1
scikit-learn==1.6.1
scipy==1.15.1
scmrepo==3.3.10
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
sqltrie==0.11.1
stack-data==0.6.3
starlette==0.47.1
tabulate==0.9.0
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treelib==1.7.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.3.0
userpath==1.9.2
uvicorn==0.35.0
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.2
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.2
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0